Move crash handler to a function
