+++ source/common.h
@@ -15,5 +15,6 @@ PyObject *aws_py_get_cpu_group_count(PyObject *self, PyObject *args);
 PyObject *aws_py_get_cpu_count_for_group(PyObject *self, PyObject *args);
 
 PyObject *aws_py_thread_join_all_managed(PyObject *self, PyObject *args);
+PyObject *aws_py_install_crash_handler(PyObject *self, PyObject *args);
 
 #endif /* AWS_CRT_PYTHON_COMMON_H */
